added cuddle-counter. maturationTimes are calculated more precisely to avoid rounding errors with high multipliers.

+++ ARKBreedingStats/BreedingInfo.cs
@@ -39,9 +39,9 @@ namespace ARKBreedingStats
                     switch (k)
                     {
                         default:
-                        case 0: t1 = breeding.pregnancyTimeAdjusted == 0 ? breeding.incubationTimeAdjusted : breeding.pregnancyTimeAdjusted; totalTime = t1; break;
+                        case 0: t1 = (int)(breeding.pregnancyTimeAdjusted == 0 ? breeding.incubationTimeAdjusted : breeding.pregnancyTimeAdjusted); totalTime = t1; break;
                         case 1: t1 = (int)(.1f * breeding.maturationTimeAdjusted); totalTime = t1; break;
-                        case 2: t1 = breeding.maturationTimeAdjusted; totalTime = breeding.pregnancyTimeAdjusted + breeding.incubationTimeAdjusted + breeding.maturationTimeAdjusted; break;
+                        case 2: t1 = (int)breeding.maturationTimeAdjusted; totalTime = (int)(breeding.pregnancyTimeAdjusted + breeding.incubationTimeAdjusted + breeding.maturationTimeAdjusted); break;
                     }
 
                     string[] subitems = new string[] { rowNames[k],
@@ -55,8 +55,8 @@ namespace ARKBreedingStats
 
                 breedingInfo += "\n";
 
-                TimeSpan incubation = new TimeSpan(0, 0, breeding.pregnancyTimeAdjusted + breeding.incubationTimeAdjusted);
-                TimeSpan growing = new TimeSpan(0, 0, breeding.maturationTimeAdjusted);
+                TimeSpan incubation = new TimeSpan(0, 0, (int)(breeding.pregnancyTimeAdjusted + breeding.incubationTimeAdjusted));
+                TimeSpan growing = new TimeSpan(0, 0, (int)breeding.maturationTimeAdjusted);
                 buttonHatching.Text = firstTime;
 
                 // further info
@@ -68,7 +68,7 @@ namespace ARKBreedingStats
                 if (breeding.eggTempMin > 0 && breeding.matingCooldownMinAdjusted > 0)
                     breedingInfo += "\n\n";
                 if (breeding.matingCooldownMinAdjusted > 0)
-                    breedingInfo += "Time until next mating is possible:\n" + new TimeSpan(0, 0, breeding.matingCooldownMinAdjusted).ToString("d':'hh':'mm") + " - " + new TimeSpan(0, 0, breeding.matingCooldownMaxAdjusted).ToString("d':'hh':'mm");
+                    breedingInfo += "Time until next mating is possible:\n" + new TimeSpan(0, 0, (int)breeding.matingCooldownMinAdjusted).ToString("d':'hh':'mm") + " - " + new TimeSpan(0, 0, (int)breeding.matingCooldownMaxAdjusted).ToString("d':'hh':'mm");
                 labelBreedingInfos.Text = breedingInfo;
             }
         }
